+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Add `extraJdkModules` option to `customJdk`
 - Add `wrap` option to `mkBabashka`
 - Check that the `main-ns` has a `:gen-class` in `mkCljBin` (see #39, thanks
   @slimslenderslacks)